requirements.txt
pytest==8.3.5
uiautomator2==3.2.10
pyyaml==6.0.2
pytest-xdist==3.6.1
pytest-html==4.1.1 